Restructured files and basic setup for new dashboard components
