+++ src/components/Dashboard/Dashboard.js
@@ -1,27 +1,32 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useLocation } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { SearchBar, Filters } from "./SearchBar&Filters";
-import DashboardNav from "./DashboardNav";
-import SuggestedJobs from "./SuggestedJobs.js";
+import PrivateRoute from "../../utils/PrivateRoute.js"
+import DashboardNav from "./DashboardNav.js";
+import SideBar from "./SideBar.js";
+import JobDetails from "./Jobs/JobDetails.js";
+import SavedJobs from "./Jobs/SavedJobs.js";
+import SuggestedJobs from "./Jobs/SuggestedJobs.js";
+import ViewedJobs from "./Jobs/ViewedJobs.js";
 
 
 function Dashboard() {
     const location = useLocation();
 
     return (
-        <>
-            <div style={{display: 'flex'}}>
-                <div className="dashboard-SideBar">
-                    <SearchBar />
-                    <Filters />
-                </div>
-                <div style={{display: 'flex', flexWrap: 'wrap', width: '100%', }}>
-                    <DashboardNav />
-                    <SuggestedJobs />
-                </div>
+    <>
+        <div style={{display: 'flex'}}>
+            <SideBar />
+
+            <div style={{display: 'flex', flexWrap: 'wrap', width: '100%', }}>
+                <DashboardNav />
+                <PrivateRoute path='/Dashboard' component={SuggestedJobs} />
+                <PrivateRoute path='/Dashboard/Job/:id' component={JobDetails} />
+                <PrivateRoute path='/Dashboard/Saved' component={SavedJobs} />
+                <PrivateRoute path='/Dashboard/Viewed' component={ViewedJobs} />
             </div>
-        </>
+        </div>
+    </>
     )
 }
