+++ src/components/Dashboard/Jobs/SavedJobs.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SavedJobs() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
