+++ src/components/Dashboard/Jobs/ViewedJobs.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+
+
+
+
+
+
+export default function ViewedJobs() {
+
+
+    return (
+        <div>
+            
+        </div>
+    )
+}
